+++ riscv/riscv.mk.in
@@ -39,6 +39,7 @@ riscv_precompiled_hdrs = \
 	insn_template.h \
 
 riscv_srcs = \
+	isa_parser.cc \
 	processor.cc \
 	execute.cc \
 	dts.cc \
